refactor(web-app): consolidate job status badge component
- Moved JobStatusBadge component to a shared location for reuse across job tables.
- Removed the obsolete JobStatusBadge implementation from the agent-specific job table.
- Updated imports to reflect the new location of the JobStatusBadge component.
