+++ web-app/src/app/app/jobs/jobs-table.tsx
@@ -4,6 +4,7 @@ import { ColumnDef, createColumnHelper } from "@tanstack/react-table";
 import { useFormatter, useTranslations } from "next-intl";
 import React from "react";
 
+import JobStatusBadge from "@/app/jobs/job-status-badge";
 import { DataTable, DataTableColumnHeader } from "@/components/data-table";
 import { JobWithRelations } from "@/lib/db/services/job.service";
 
@@ -33,34 +34,6 @@ export default function JobsTable({ jobs }: JobsTableProps) {
       showPagination
       defaultSort={defaultSort}
     />
-  );
-}
-
-// Job status badge component
-function JobStatusBadge({ status }: { status: string }) {
-  const getStatusColor = () => {
-    switch (status) {
-      case "COMPLETED":
-        return "bg-green-100 text-green-800";
-      case "PROCESSING":
-        return "bg-blue-100 text-blue-800";
-      case "FAILED":
-        return "bg-red-100 text-red-800";
-      case "PAYMENT_PENDING":
-        return "bg-yellow-100 text-yellow-800";
-      case "PAYMENT_FAILED":
-        return "bg-red-100 text-red-800";
-      default:
-        return "bg-gray-100 text-gray-800";
-    }
-  };
-
-  return (
-    <span
-      className={`inline-flex items-center rounded-full px-2.5 py-0.5 text-xs font-medium ${getStatusColor()}`}
-    >
-      {status}
-    </span>
   );
 }
